dded BH functionality
I introduced a few things in this commmit
I added the guzzle http library, and created a facade for it so it
can be globally reached by using Guzzle:: you have to create a service provider then create a facade for that service provider.

I also created a blue hornet service provider, if you have params for something that you want to depenency inject you need to create a service provider to be able to bind params for the resolution.