90% of User Crud
Sorry i got in the zone and did not chunk this but, i fixed some issues in the esp view, and separated repeated logic in angular into its own file.  Basically tok my registration screen and moved it.  need to get rid of old routes when done.  still need to handle errors on edit, and role crud.